Initialize document store when DocumentStoreHolder is created to avoid object lifecycle management issues
