--- src/IdentityServer4.RavenDB.Storage/DocumentStoreHolder/DocumentStoreHolderBase.cs
@@ -1,24 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-using Raven.Client.Documents;
-using Raven.Client.Documents.Session;
-
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
-namespace IdentityServer4.RavenDB.Storage.DocumentStoreHolder
-{
-    internal abstract class DocumentStoreHolderBase
-    {
-        public IDocumentStore DocumentStore;
-
-        protected DocumentStoreHolderBase(IDocumentStore documentStore)
-        {
-            DocumentStore = documentStore;
-        }
-
-        public IAsyncDocumentSession OpenAsyncSession() => DocumentStore.OpenAsyncSession();
-
-        public virtual void Dispose()
-        {
-            DocumentStore?.Dispose();
-        }
-    }
-}
