+++ src/IdentityServer4.RavenDB.Storage/DocumentStoreHolder/OperationalDocumentStoreHolder.cs
@@ -1,14 +1,28 @@
 ﻿using System;
-using System.Runtime.CompilerServices;
+using IdentityServer4.RavenDB.Storage.Helpers;
+using IdentityServer4.RavenDB.Storage.Options;
 using Raven.Client.Documents;
+using Raven.Client.Documents.Session;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.DocumentStoreHolder
 {
-    internal class OperationalDocumentStoreHolder : DocumentStoreHolderBase, IDisposable
+    internal class OperationalDocumentStoreHolder : IDisposable
     {
-        public OperationalDocumentStoreHolder(IDocumentStore documentStore) : base(documentStore)
+        private readonly IDocumentStore _documentStore;
+        
+        public OperationalDocumentStoreHolder(RavenDbOperationalStoreOptions options)
         {
+            _documentStore = DocumentStoreHelper.InitializeDocumentStore(options.ConfigureDocumentStore);
+            IndexHelper.ExecuteOperationalStoreIndexes(_documentStore);
+        }
+        
+        public IDocumentStore IntegrationTest_GetDocumentStore() => _documentStore;
+        
+        public IAsyncDocumentSession OpenAsyncSession() => _documentStore.OpenAsyncSession();
+
+        public void Dispose()
+        {
+            _documentStore?.Dispose();
         }
     }
 }
