+++ src/IdentityServer4.RavenDB.Storage/Entities/ApiScope.cs
@@ -1,6 +1,4 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
+﻿
 namespace IdentityServer4.RavenDB.Storage.Entities
 {
     internal class ApiScope : Resource
