+++ src/IdentityServer4.RavenDB.Storage/Entities/ClientClaim.cs
@@ -1,7 +1,5 @@
-﻿using System.Runtime.CompilerServices;
-using System.Security.Claims;
+﻿using System.Security.Claims;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Entities
 {
     internal class ClientClaim
