+++ src/IdentityServer4.RavenDB.Storage/Entities/DeviceFlowCode.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Entities
 {
     internal class DeviceFlowCode
