+++ src/IdentityServer4.RavenDB.Storage/Entities/Resource.cs
@@ -1,8 +1,6 @@
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Entities
 {
     /// <summary>
