+++ src/IdentityServer4.RavenDB.Storage/Helpers/DocumentStoreHelper.cs
@@ -18,9 +18,14 @@ namespace IdentityServer4.RavenDB.Storage.Helpers
                 throw new InvalidOperationException($"{nameof(documentStore.Database)} must be provided when setting up Identity Server configuration and operational RavenDb stores." );
             }
 
+            if (urls == null)
+            {
+                throw new InvalidOperationException($"{nameof(documentStore.Urls)} must be provided when setting up Identity Server configuration and operational RavenDb stores." );
+            }
+
             if (urls.Length == 0)
             {
-                throw new InvalidOperationException($"{nameof(documentStore.Urls)} must be provided when setting up Identity Server configuration and operational RavenDb stores.");
+                throw new InvalidOperationException($"{nameof(documentStore.Urls)} cannot be empty when setting up Identity Server configuration and operational RavenDb stores.");
             }
 
             if (documentStore.Certificate != null)
