+++ src/IdentityServer4.RavenDB.Storage/Helpers/IndexHelper.cs
@@ -1,10 +1,8 @@
 ﻿using System.Collections.Generic;
-using System.Runtime.CompilerServices;
 using IdentityServer4.RavenDB.Storage.Indexes;
 using Raven.Client.Documents;
 using Raven.Client.Documents.Indexes;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Helpers
 {
     internal static class IndexHelper
@@ -37,13 +35,8 @@ namespace IdentityServer4.RavenDB.Storage.Helpers
         {
             foreach (var index in indexes)
             { 
-                ExecuteIndex(store, index, store.Database);
+                store.ExecuteIndex(index, store.Database);
             }
-        }
-        
-        private static void ExecuteIndex(IDocumentStore store, AbstractIndexCreationTask index, string databaseName)
-        {
-            store.ExecuteIndex(index, databaseName);
         }
     }
 }
