+++ src/IdentityServer4.RavenDB.Storage/IdentityServer4.RavenDB.Storage.csproj
@@ -28,4 +28,10 @@
     <PackageReference Include="RavenDB.Client" Version="4.2.105" />
   </ItemGroup>
 
+  <ItemGroup>
+    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+      <_Parameter1>$(AssemblyName).Tests</_Parameter1>
+    </AssemblyAttribute>
+  </ItemGroup>
+
 </Project>
