+++ src/IdentityServer4.RavenDB.Storage/Indexes/ApiResourceIndex.cs
@@ -1,10 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using IdentityServer4.RavenDB.Storage.Entities;
 using Raven.Client.Documents.Indexes;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Indexes
 {
     internal class ApiResourceIndex : AbstractIndexCreationTask<ApiResource, ApiResourceIndex.Result>
