+++ src/IdentityServer4.RavenDB.Storage/Indexes/ClientIndex.cs
@@ -1,10 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using IdentityServer4.RavenDB.Storage.Entities;
 using Raven.Client.Documents.Indexes;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Indexes
 {
     internal class ClientIndex : AbstractIndexCreationTask<Client, ClientIndex.Result>
