+++ src/IdentityServer4.RavenDB.Storage/Mappers/ApiResourceMappers.cs
@@ -1,8 +1,6 @@
-﻿using System.Runtime.CompilerServices;
-using AutoMapper;
+﻿using AutoMapper;
 using IdentityServer4.RavenDB.Storage.Entities;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Mappers
 {
     internal static class ApiResourceMappers
