+++ src/IdentityServer4.RavenDB.Storage/Mappers/ClientMappers.cs
@@ -1,7 +1,5 @@
-﻿using System.Runtime.CompilerServices;
-using AutoMapper;
+﻿using AutoMapper;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Mappers
 {
     internal static class ClientMappers
