+++ src/IdentityServer4.RavenDB.Storage/Mappers/PersistedGrantMapperProfile.cs
@@ -1,8 +1,6 @@
-﻿using System.Runtime.CompilerServices;
-using AutoMapper;
+﻿using AutoMapper;
 using IdentityServer4.RavenDB.Storage.Helpers;
 
-[assembly: InternalsVisibleTo("IdentityServer4.RavenDB.IntegrationTests")]
 namespace IdentityServer4.RavenDB.Storage.Mappers
 {
     /// <summary>
