+++ test/IdentityServer4.RavenDB.Storage.Tests/DocumentStoreHolderTests/certificate.txt
