+++ test/IdentityServer4.RavenDB.Storage.Tests/IdentityServer4.RavenDB.Storage.Tests.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>IdentityServer4.RavenDB.Storage.Tests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
