+++ test/IdentityServer4.RavenDB.Storage.Tests/StoresTests/ClientStoreTests.cs
@@ -2,23 +2,21 @@
 using System.Threading.Tasks;
 using FluentAssertions;
 using IdentityServer4.Models;
-using IdentityServer4.RavenDB.Storage.DocumentStoreHolder;
 using IdentityServer4.RavenDB.Storage.Helpers;
-using IdentityServer4.RavenDB.Storage.Indexes;
 using IdentityServer4.RavenDB.Storage.Mappers;
 using IdentityServer4.RavenDB.Storage.Stores;
 using Xunit;
 using Xunit.Sdk;
 
-namespace IdentityServer4.RavenDB.IntegrationTests.StoresTests
+namespace IdentityServer4.RavenDB.Storage.Tests.StoresTests
 {
     public class ClientStoreTests : IntegrationTestBase
     {
         [Fact]
         public async Task FindClientByIdAsync_WhenClientDoesNotExist_ExpectNull()
         {
-            var storeHolder = await GetOperationalDocumentStoreHolder_AndExecuteClientIndex();
-            WaitForIndexing(storeHolder.DocumentStore);
+            var storeHolder = GetConfigurationDocumentStoreHolder();
+            WaitForIndexing(storeHolder.IntegrationTest_GetDocumentStore());
             
             var store = new ClientStore(storeHolder, FakeLogger<ClientStore>.Create());
             var client = await store.FindClientByIdAsync(Guid.NewGuid().ToString());
@@ -28,8 +26,8 @@ namespace IdentityServer4.RavenDB.IntegrationTests.StoresTests
         [Fact]
         public async Task FindClientByIdAsync_WhenClientExists_ExpectClientReturned()
         {
-            var storeHolder = await GetOperationalDocumentStoreHolder_AndExecuteClientIndex();
-            WaitForIndexing(storeHolder.DocumentStore);
+            var storeHolder = GetConfigurationDocumentStoreHolder();
+            WaitForIndexing(storeHolder.IntegrationTest_GetDocumentStore());
             
             var testClient = new Client
             {
@@ -43,7 +41,7 @@ namespace IdentityServer4.RavenDB.IntegrationTests.StoresTests
                 await session.SaveChangesAsync();
             }
 
-            WaitForIndexing(storeHolder.DocumentStore);
+            WaitForIndexing(storeHolder.IntegrationTest_GetDocumentStore());
 
             var store = new ClientStore(storeHolder, FakeLogger<ClientStore>.Create());
             var client = await store.FindClientByIdAsync(testClient.ClientId);
@@ -54,7 +52,7 @@ namespace IdentityServer4.RavenDB.IntegrationTests.StoresTests
         [Fact]
         public async Task FindClientByIdAsync_WhenClientExistsWithCollections_ExpectClientReturnedCollections()
         {
-            var storeHolder = await GetOperationalDocumentStoreHolder_AndExecuteClientIndex();
+            var storeHolder = GetConfigurationDocumentStoreHolder();
 
             var testClient = new Client
             {
@@ -77,7 +75,7 @@ namespace IdentityServer4.RavenDB.IntegrationTests.StoresTests
                 await session.SaveChangesAsync();
             }
 
-            WaitForIndexing(storeHolder.DocumentStore);
+            WaitForIndexing(storeHolder.IntegrationTest_GetDocumentStore());
 
             var store = new ClientStore(storeHolder, FakeLogger<ClientStore>.Create());
             var client = await store.FindClientByIdAsync(testClient.ClientId);
@@ -88,7 +86,7 @@ namespace IdentityServer4.RavenDB.IntegrationTests.StoresTests
         [Fact]
         public async Task FindClientByIdAsync_WhenClientsExistWithManyCollections_ExpectClientReturnedInUnderFiveSeconds()
         {
-            var storeHolder = await GetOperationalDocumentStoreHolder_AndExecuteClientIndex();
+            var storeHolder = GetConfigurationDocumentStoreHolder();
 
             var testClient = new Client
             {
@@ -126,7 +124,7 @@ namespace IdentityServer4.RavenDB.IntegrationTests.StoresTests
                 await session.SaveChangesAsync();
             }
 
-            WaitForIndexing(storeHolder.DocumentStore);
+            WaitForIndexing(storeHolder.IntegrationTest_GetDocumentStore());
 
             var store = new ClientStore(storeHolder, FakeLogger<ClientStore>.Create());
 
@@ -142,13 +140,6 @@ namespace IdentityServer4.RavenDB.IntegrationTests.StoresTests
             {
                 throw new TestTimeoutException(timeout);
             }
-        }
-        
-        private async Task<ConfigurationDocumentStoreHolder> GetOperationalDocumentStoreHolder_AndExecuteClientIndex()
-        {
-            var storeHolder = GetConfigurationDocumentStoreHolder();
-            await ExecuteIndex(storeHolder.DocumentStore, new ClientIndex());
-            return storeHolder;
         }
     }
 }
